Code CleanUp Done
