+++ src/Types/CityTypes.ts
@@ -1,0 +1,9 @@
+export interface CityType {
+  id: number;
+  name: string;
+  region: string;
+  country: string;
+  lat: number;
+  lon: number;
+  url: string;
+}
